rename amm to xyk
